backend/app/workers/clip_processor.py
"""
Celery worker for processing video clip extraction.
Requirements: 4.2, 6.2
"""
from celery import Task
from app.core.celery_app import celery_app
from app.core.storage import storage_service
from app.db.base import SessionLocal
from app.models.clip import Clip
from datetime import datetime
import cv2
import numpy as np
import io
import logging
from uuid import UUID
from PIL import Image

logger = logging.getLogger(__name__)


class ClipProcessingTask(Task):
    """Base task with database session management."""
    
    def __call__(self, *args, **kwargs):
        """Execute task with database session."""
        return self.run(*args, **kwargs)


@celery_app.task(bind=True, base=ClipProcessingTask, name="process_clip")
def process_clip_task(
    self,
    clip_id: str,
    user_id: str,
    start_ts: str,
    duration_ms: int,
    labels: list = None
):
    """
    Process a video clip extraction job.
    
    Args:
        clip_id: UUID of the clip record
        user_id: UUID of the user
        start_ts: ISO format timestamp of clip start
        duration_ms: Duration of clip in milliseconds
        labels: Optional list of labels for the clip
    """
    db = SessionLocal()
    try:
        logger.info(f"Processing clip {clip_id}")
        
        # Update clip status to processing
        clip = db.query(Clip).filter(Clip.id == UUID(clip_id)).first()
        if not clip:
            logger.error(f"Clip {clip_id} not found")
            return {"status": "error", "message": "Clip not found"}
        
        # For now, create a placeholder video file
        # In production, this would extract actual video segments
        # from a video buffer or recording system
        
        # Generate a simple preview frame
        preview_image = Image.new('RGB', (640, 480), color=(50, 50, 150))
        preview_output = io.BytesIO()
        preview_image.save(preview_output, format='JPEG', quality=85)
        preview_data = preview_output.getvalue()
        
        # Upload preview to S3
        timestamp_str = datetime.fromisoformat(start_ts).strftime("%Y%m%d_%H%M%S")
        preview_key = f"clips/{user_id}/{timestamp_str}_preview.jpg"
        preview_uri = storage_service.upload_file(
            file_data=preview_data,
            object_key=preview_key,
            content_type="image/jpeg"
        )
        
        # Create a placeholder video file (1 second of black frames)
        # In production, this would be actual video extraction
        video_key = f"clips/{user_id}/{timestamp_str}_clip.mp4"
        
        # For now, just create a minimal MP4 placeholder
        # In production, use ffmpeg or similar to extract actual video
        placeholder_video = b"placeholder_video_data"
        video_uri = storage_service.upload_file(
            file_data=placeholder_video,
            object_key=video_key,
            content_type="video/mp4"
        )
        
        # Update clip record with S3 URIs
        clip.s3_uri = video_uri
        clip.preview_png = preview_uri
        if labels:
            clip.labels = labels
        
        db.commit()
        db.refresh(clip)
        
        logger.info(f"Clip {clip_id} processed successfully")
        
        return {
            "status": "completed",
            "clip_id": clip_id,
            "s3_uri": video_uri,
            "preview_uri": preview_uri
        }
        
    except Exception as e:
        logger.error(f"Error processing clip {clip_id}: {e}")
        db.rollback()
        return {"status": "error", "message": str(e)}
    finally:
        db.close()
