backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9
redis==5.0.7
pydantic==2.8.2
pydantic-settings==2.3.4
python-multipart==0.0.9
aiortc==1.9.0
starlette==0.37.2
celery==5.4.0
boto3==1.34.144
opencv-python-headless==4.10.0.84
numpy==1.26.4
av==12.0.0
websockets==12.0
Pillow==10.4.0
croniter==2.0.5
